+++ packages/overlay/test/overlay-trigger-click.test.ts
@@ -28,7 +28,7 @@ import {
 import '@spectrum-web-components/overlay/overlay-trigger.js';
 import { spy } from 'sinon';
 import { ActionButton } from '@spectrum-web-components/action-button';
-import { fixture, isInteractive } from '../../../test/testing-helpers.js';
+import { fixture, isOnTopLayer } from '../../../test/testing-helpers.js';
 
 describe('Overlay Trigger - Click', () => {
     it('displays `click` declaratively', async () => {
@@ -68,18 +68,25 @@ describe('Overlay Trigger - Click', () => {
         });
     });
     describe('closes on scroll', () => {
-        afterEach(() => {
+        afterEach(async () => {
             if (document.scrollingElement) {
                 document.scrollingElement.scrollTop = 0;
             }
+            await waitUntil(() => {
+                if (document.scrollingElement) {
+                    return document.scrollingElement.scrollTop === 0;
+                }
+                return true;
+            });
         });
         (['click', 'replace', 'inline'] as TriggerInteractionsV1[]).map(
             (interaction) => {
-                it(`closes "${interaction}" overlay on scroll`, async () => {
+                it(`closes "${interaction}" overlay on scroll`, async function () {
                     const el = await fixture<OverlayTrigger>(html`
                         <overlay-trigger
-                            placement="right-start"
+                            placement="right"
                             type=${interaction}
+                            content="click"
                         >
                             <sp-action-button
                                 slot="trigger"
@@ -87,7 +94,9 @@ describe('Overlay Trigger - Click', () => {
                             >
                                 <sp-icon-magnify slot="icon"></sp-icon-magnify>
                             </sp-action-button>
-                            <sp-popover slot="click-content" tip></sp-popover>
+                            <sp-popover slot="click-content" tip>
+                                Content
+                            </sp-popover>
                         </overlay-trigger>
                     `);
                     await nextFrame();
@@ -96,20 +105,27 @@ describe('Overlay Trigger - Click', () => {
 
                     await elementUpdated(el);
                     const opened = oneEvent(el, 'sp-opened');
-                    el.open = 'click';
+                    const trigger = el.querySelector(
+                        'sp-action-button'
+                    ) as HTMLElement;
+                    trigger.click();
+
                     await opened;
 
                     expect(el.open).to.equal('click');
-                    expect(await isInteractive(popover)).to.be.true;
+
+                    expect(await isOnTopLayer(popover)).to.be.true;
 
                     const closed = oneEvent(el, 'sp-closed');
                     if (document.scrollingElement) {
                         document.scrollingElement.scrollTop = 100;
                     }
+
                     await closed;
 
                     expect(el.open).to.be.undefined;
-                    expect(await isInteractive(popover)).to.be.false;
+
+                    expect(await isOnTopLayer(popover)).to.be.false;
                 });
             }
         );