+++ web-test-runner.config.js
@@ -16,7 +16,7 @@ import {
 import { sendMousePlugin } from './test/plugins/send-mouse-plugin.js';
 import {
     chromium,
-    chromiumWithFlags,
+    chromiumWithMemoryTooling,
     configuredVisualRegressionPlugin,
     firefox,
     packages,
@@ -54,6 +54,13 @@ export default {
                 }
             },
         },
+        {
+            name: 'measureUserAgentSpecificMemory-plugin',
+            transform(context) {
+                context.set('Cross-Origin-Opener-Policy', 'same-origin');
+                context.set('Cross-Origin-Embedder-Policy', 'credentialless');
+            },
+        },
     ],
     mimeTypes: {
         '**/*.json': 'js',
@@ -63,8 +70,6 @@ export default {
     },
     http2: true,
     protocol: 'https:',
-    concurrency: 4,
-    concurrentBrowsers: 1,
     testsFinishTimeout: 60000,
     coverageConfig: {
         report: true,
@@ -73,18 +78,14 @@ export default {
             'packages/*/stories/*',
             'packages/icons-ui/**',
             'packages/icons-workflow/**',
-            // The following file is no longer used in Chrome where coverage is calculated.
             'test/**',
             '**/test/**',
             'tools/*/stories/*',
-            'tools/shared/src/focus-visible.*',
             'tools/styles/**',
             '**/node_modules/**',
-            // The following are WIP removals for the Overlay API update
-            '**/ActiveOverlay.*',
-            '**/overlay-stack.*',
-            '**/overlay-utils.*',
-            '**/OverlayPopover.*',
+            // The following files are not used in Chrome where coverage is calculated.
+            '**/OverlayNoPopover.*',
+            'tools/shared/src/focus-visible.*',
         ],
         threshold: {
             statements: 98.5,
@@ -95,7 +96,8 @@ export default {
     },
     testFramework: {
         config: {
-            timeout: 3000,
+            timeout: 5000,
+            retries: 1,
         },
     },
     groups: [
@@ -131,9 +133,9 @@ export default {
                 'packages/split-button/test/*.test.js',
                 'packages/tooltip/test/*.test.js',
             ],
-            browsers: [chromium, chromiumWithFlags, firefox, webkit],
+            browsers: [chromium, firefox, webkit],
         },
     ],
     group: 'unit',
-    browsers: [chromium, firefox, webkit],
+    browsers: [firefox, chromiumWithMemoryTooling, webkit],
 };